Add RTC Battery Charging
Previous impl. was setting a value for runtime config in hope
that offspot-runtime would support it and enable the charger.

While offspot-runtime may support it, given it is enabled in the firmware
it would require a restart which would be unwanted in our provisioning
scenario.

This changes it it so charger is enabled from next boot.

Other changes:
- commented OffspotYAMLStep as it's now doing nothing
- Moved boot order step last as this should only be done if all other
steps succeeded.

Fixes https://github.com/offspot/base-image/issues/74